+++ src/main/java/fr/maxlego08/essentials/commands/commands/teleport/random/CommandTeleportRandom.java
@@ -1,4 +1,4 @@
-package fr.maxlego08.essentials.commands.commands.teleport;
+package fr.maxlego08.essentials.commands.commands.teleport.random;
 
 import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.CommandResultType;
@@ -21,7 +21,7 @@ public class CommandTeleportRandom extends VCommand {
     protected CommandResultType perform(EssentialsPlugin plugin) {
 
         TeleportationModule module = plugin.getModuleManager().getModule(TeleportationModule.class);
-        module.randomTeleport(player);
+        module.randomTeleport(player, player.getWorld());
 
         return CommandResultType.SUCCESS;
     }