+++ src/main/java/fr/maxlego08/essentials/module/modules/TeleportationModule.java
@@ -26,6 +26,7 @@ public class TeleportationModule extends ZModule {
     private final List<TeleportPermission> teleportDelayPermissions = new ArrayList<>();
     private final List<TeleportPermission> teleportProtections = new ArrayList<>();
     private final List<String> blacklistBiomes = new ArrayList<>();
+    private final List<RandomTeleportWorld> rtpWorlds = new ArrayList<>();
     private boolean teleportSafety;
     private boolean teleportToCenter;
     private int teleportDelay;
@@ -35,7 +36,7 @@ public class TeleportationModule extends ZModule {
     private boolean openConfirmInventoryForTpa;
     private boolean openConfirmInventoryForTpaHere;
     private int maxRtpAttempts;
-    private final List<RandomTeleportWorld> rtpWorlds = new ArrayList<>();
+    private boolean enableRandomTeleportSearchLogMessage;
     @NonLoadable
     private Map<String, RandomTeleportWorld> rtpWorldMap = new HashMap<>();
 
@@ -102,51 +103,59 @@ public class TeleportationModule extends ZModule {
         this.plugin.getInventoryManager().openInventory(player, this.plugin, "confirm_request_here_inventory");
     }
 
-    public void randomTeleport(Player player) {
-        RandomTeleportWorld configuration = this.rtpWorldMap.get(player.getWorld().getName());
+    public void randomTeleport(Player player, World world) {
+
+        RandomTeleportWorld configuration = this.rtpWorldMap.get(world.getName());
         if (configuration == null) {
-            message(player, Message.COMMAND_RANDOM_TP_ERROR);
+            message(player, Message.COMMAND_RANDOM_TP_CONFIGURATION_NOT_FOUND, "world", world.getName());
             return;
         }
 
-        World world = player.getWorld();
         this.randomTeleport(player, world, configuration.centerX(), configuration.centerZ(), configuration.radiusX(), configuration.radiusZ());
     }
 
 
     public void randomTeleport(Player player, World world, int centerX, int centerZ, int rangeX, int rangeZ) {
+        this.debug("Starting random teleport for player " + player.getName());
         message(player, Message.TELEPORT_RANDOM_START);
         getRandomSurfaceLocation(world, centerX, centerZ, rangeX, rangeZ, this.maxRtpAttempts).thenAccept(randomLocation -> {
+            this.debug("Random location found: " + randomLocation);
             if (randomLocation != null) {
                 User user = this.getUser(player);
                 user.teleport(randomLocation, Message.TELEPORT_MESSAGE_RANDOM, Message.TELEPORT_SUCCESS_RANDOM);
             } else {
+                this.debug("Failed to find random location");
                 message(player, Message.COMMAND_RANDOM_TP_ERROR);
             }
         });
     }
 
     private CompletableFuture<Location> getRandomSurfaceLocation(World world, int centerX, int centerZ, int rangeX, int rangeZ, int attempts) {
+        this.debug("Starting random surface location search for world " + world.getName());
         CompletableFuture<Location> future = new CompletableFuture<>();
 
         if (attempts > 0) {
 
             randomLocation(world, centerX, centerZ, rangeX, rangeZ).thenAccept(location -> {
+                this.debug("Random location generated: " + location);
                 if (isValidLocation(location)) {
                     future.complete(location);
                 } else {
+                    this.debug("Random location not valid");
                     getRandomSurfaceLocation(world, centerX, centerZ, rangeX, rangeZ, attempts - 1).thenAccept(future::complete);
                 }
             });
 
         } else {
-            future.complete(new Location(world, centerX, 256, centerZ));
+            this.debug("Failed to find random surface location, using default location");
+            future.complete(null);
         }
 
         return future;
     }
 
     private CompletableFuture<Location> randomLocation(World world, int centerX, int centerZ, int rangeX, int rangeZ) {
+        this.debug("Generating random location for world " + world.getName());
         CompletableFuture<Location> future = new CompletableFuture<>();
 
         int x = centerX + (int) (Math.random() * (2 * rangeX + 1)) - rangeX;
@@ -155,8 +164,10 @@ public class TeleportationModule extends ZModule {
 
         world.getChunkAtAsync(x, z).thenAccept(chunk -> {
             Location location = new Location(world, x + 0.5, y, z + 0.5, 360 * random.nextFloat() - 180, 0);
+            this.debug("Generated random location: " + location);
             this.plugin.getScheduler().runAtLocation(location, wrappedTask -> {
                 location.setY(World.Environment.NETHER == world.getEnvironment() ? getNetherYAt(location) : world.getHighestBlockYAt(location));
+                this.debug("Final location determined: " + location);
                 future.complete(location);
             });
         });
@@ -165,7 +176,10 @@ public class TeleportationModule extends ZModule {
     }
 
     private boolean isValidLocation(Location location) {
-        return this.blacklistBiomes.stream().noneMatch(b -> b.equalsIgnoreCase(location.getBlock().getBiome().name())) && location.getBlock().getType().isSolid() && location.add(0, 1, 0).getBlock().getType().isAir() && location.add(0, 2, 0).getBlock().getType().isAir();
+        return this.blacklistBiomes.stream().noneMatch(b -> b.equalsIgnoreCase(location.getBlock().getBiome().name())) // 
+                && location.getBlock().getType().isSolid() // 
+                && location.add(0, 1, 0).getBlock().getType().isAir() // 
+                && location.add(0, 2, 0).getBlock().getType().isAir();
     }
 
     private int getNetherYAt(final Location location) {
@@ -180,5 +194,11 @@ public class TeleportationModule extends ZModule {
     private boolean isBlockUnsafe(World world, int x, int y, int z) {
         Material blockType = world.getBlockAt(x, y, z).getType();
         return !blockType.isSolid() || blockType == Material.LAVA || blockType == Material.FIRE;
+    }
+
+    private void debug(String message) {
+        if (this.enableRandomTeleportSearchLogMessage) {
+            this.plugin.getLogger().info(message);
+        }
     }
 }